backend/pyproject.toml
[project]
name = "project-kisan-backend"
version = "0.1.0"
description = "AI-Powered Agricultural Assistant - Hackathon MVP"
authors = [
    {name = "CodeKheti.ai Team", email = "team@codekheti.ai"},
]
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6", # For file uploads
    "google-cloud-aiplatform>=1.38.0", # Vertex AI
    "google-cloud-speech>=2.22.0", # Speech APIs
    "google-cloud-texttospeech>=2.27.0", # Text-to-Speech API
    "google-cloud-translate>=3.12.0", # Translation API
    "google-cloud-storage>=2.10.0", # File storage
    "google-cloud-firestore>=2.14.0", # Firestore database
    "requests>=2.31.0", # External APIs
    "python-dotenv>=1.0.0", # Environment variables
    "pydantic>=2.5.0",
    "pydantic-settings>=2.0.0", # For BaseSettings
    "pillow>=10.1.0", # Image processing
    "aiohttp>=3.12.14",
    "structlog>=24.5.0", # Structured logging
]
readme = "README.md"
requires-python = ">=3.13"

[dependency-groups]
dev = [
    "ruff>=0.11.13",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py313"
exclude = [
    ".venv",
    "venv", 
    "__pycache__",
    "app/ai/prompts/*",
    "app/ai/prompts/**/*.py",
    "tests/*",
    "tests/**/*.py",
    "scripts/*",
    "scripts/**/*.py",
]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings  
    "F",  # Pyflakes
    "I",  # isort (import sorting)
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = ["B008", "B904"]
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.isort]
force-wrap-aliases = true
combine-as-imports = true
